requirements.txt
numpy==1.22.3
scipy==1.9.3
scikit-learn==1.0.2
scikit-image==0.19.2
pyod==1.0.6
matplotlib==3.5.3
opencv-python==4.6.0.66
jpype1==1.3.0
numba==0.56.3
java 8
